Cleanup / Restructuring of backend code.
The WMMA intrinsics are now encapsulated in a struct, so we can reuse
the tiling and caching code across different GPU targets, as long as we
can come up with some WMMA style matrix multiplication and the
associated load/store operations for the A, B, C, and D matrix.

Small caveat: The code assumes a wave/warp size of 32 on all GPU targets.

In addition to this, a BACKEND cmake variable was introduced to switch
between NVVM, HSA, and OneAPI Backends.