Unify LZ and other GPU code.
This is done by putting typedefs and the tile sizes in the ops files
that the tests load. Some work had to be done to cast everything when
needed. Some targets trigger "cast source and destination identical"
warnings, and there is nothing I can do to prevent this without using
implicits; and those are currently hilariously broken.