Add support for newer RX 90xx series AMD GPUs.
The signature of the WMMA intrinsics was slightly changed, and some of
the matrix load/store code had to be altered as well. The
HSAWMMAIntrinsics struct was split into an GFX11- and
GFX12WMMAIntrinsics variant. The operations code currently uses the
GFX11 operations, but switching to the newer ones is straight forward.